Major Update
Added Logic to change Total: textField from green to red and vise versa.
Established the Welcome, to say a firstName.
Established total to appear with amount.
Fixed a nullPointerException
Added a User class to start with the database.
Added Parse to the external lib.
Removed Login Activity (Currently working on it)