Bunch of New things
Able to save friends to the database now.
Once save the friend will display on the friends list.
once clicked the friendInfo will display, Fields do not load yet.
Added a layout row that we can use later to make our Friends list rows
the way we want.
Changed the color on friendsinfo so it will actually display.
Made a separate method for the addElements in friends tab. Thats where
it populates the list using a ParseQuery.
Cleaned up some old code.
Tomorrow before the final I will edit the Decimal format to display and
also clean up more code and add comments to everything so it can be read
smoother.